+++ c/options.h
@@ -22,6 +22,7 @@ extern int8_t sort_over_radix_bits;
 extern int32_t sort_nthreads;
 extern bool fread_anonymize;
 
+int32_t normalize_nthreads(int32_t nth);
 void set_nthreads(int32_t n);
 void set_core_logger(PyObject*);
 void set_sort_insert_method_threshold(int64_t n);